fix: corrige tipos de dato en ejercicios (float -> double/int)
Auditoria de tipos detecto dos patrones incorrectos repetidos en los
ejercicios de etapa 0 y 1:

- Constantes y variables de dinero/porcentajes declaradas como float
  mientras el resto del calculo usa double (etapa-1-ejercicio-1,
  etapa-1-ejercicio-2, etapa-0-ejercicio-2). float tiene ~6-7 cifras
  significativas de precision vs ~15-16 de double, y mezclar ambos en
  la misma operacion trunca silenciosamente los literales sin dar
  error de compilacion.

- Contadores (cantidad_aprobados/cantidad_reprobados en
  etapa-0-ejercicio-1) declarados como float en vez de int, siendo
  que solo incrementan de 1 en 1 y nunca llevan decimales.

Todos los archivos se verificaron compilando con
g++ -std=c++17 -Wall antes y despues del cambio.

+++ ejercicios/etapa-0-repaso/etapa-0-ejercicio-1.cpp
@@ -5,7 +5,7 @@ int main() {
 
     const float CALIFICACION_MINIMA = 70;
     float c1, c2, c3, promedio, porcentaje_aprobados; // Declaración: Estas variables se calculan o leen desde el teclado
-    float cantidad_aprobados = 0, cantidad_reprobados = 0; // Declaración e inicialización | Contadores
+    int cantidad_aprobados = 0, cantidad_reprobados = 0; // Declaración e inicialización | Contadores
     int I, N;
     system("clear");
